Move tests outside of testing.
